Restore Net::DNS::RR->new_from_hash
There are a number of packages that expect Net::DNS::RR to have
new_from_hash() as an alternate constructor.  While recent Net::DNS
claims compatability with prior versions, for some reason this constructor
disappeared in favor of the similarly named new_hash().

This patch restores new_from_hash(), adds a test to ensure it doesn't
disappear again, and tweaks the pod tests a bit.

Thanks! :)